Change utest exit codes to avoid unnoticed errors
utest fork()s for each test case. The process exit code of the process
running the test case is used to communicate the result to the parent,
driver, process.

A UNIX exit code is a positive integer, and by convention in C error
codes are negative integers (and so also in <utest.h>). For these
reasons, in utest the sign is changed when translated from C-level
error code to process exit code (or vice versa).

The UTEST_NOT_RUN had the value -1, which was translated to the exit
code 1. Usually, it's the utest framework code running in the SUT
process that does the exit() call, so values are essentially
utest-internal. However, if the SUT application code does exit(), this
value is used by the driver process. This caused test code
exit(EXIT_FAILURE) to be mapped to NOT_RUN, which is unfortunate,
since such "not run" test cases may go unnoticed.

This patch switches the values of NOT_RUN and FAILED, so that
exit(EXIT_FAILURE) maps to UTEST_FAILED. exit(3) will still be mapped
to UTEST_NOT_RUN, but '3' is much more uncommon than '1'
(EXIT_FAILURE).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>