Add XCM relay program
Add a program 'xcmrelay' capable of acting as an XCM-level relay.

xcmrelay binds a server port, and for each inbound connection creates
an outbound connection to a configured address. Bytes (for
bytestreams) or messages (for messaging type sockets) from the
server-socket side connection will be forward as-is to the outbound
connection, and vice versa.

xcmrelay accepts socket attributes on the command-line in the same way
does the 'xcm' program.

This commit factor out parts responsible for parsing socket attribute
from the xcm program to a to a separate module, to allow reuse of this
logic.

The tools directory is restructure to accomodate the new code. Each
program now has its own subdirectory.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>