Add transport-agnostic address syntax check API
Extend <xcm_addr.h> with functionality to check if an address is
valid, and if an address is both valid and supported by the particular
build the application is linked to.

This is useful to detect errors (e.g., misconfiguration) ahead of
actual socket creation.

The XCM implementation and ABI versions are bumped to reflect the API
extension.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>