Have XCM tools accept socket attribute values on stdin
Add support in the 'xcm' and the 'xcmrelay' programs to read
binary-type socket attributes on standard input at startup. A new
command-line option '-r <attr-name>' is used for this purposes. This
option may be repeated to allow setting multiple attributes.

The use case is primarily when other programs instantiating these
tools (as opposed to a human shell operator). Now, such programs need
not create temporary files containing the attribute values.

Pipes are byte streams, and the binary socket attributes may contain
arbitrary data (including NUL characters), so the actual attribute
value must be preceded by a 32-bit length field (in network byte
order) specifying the size of the attribute (in bytes).

The xcmrelay test suite is extended to cover the new functionality.

Resolves #97.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>