Increase maximum message size for UX and UXF connections
With this change, UX (including UX-based connection on UTLS) and UXF
will make use of a xcm.max_msg_size of:

2 * <value of net.core.wmem_max> - <header margin>

The reason the net.core.wmem_max is double is due to a Linux kernel
quirk, where the actual maximum possible to set with setsockopt() is
twice that of the net.core.wmem_max administrative maximum.

The <header margin> is 128 bytes. The actual current kernel AF_UNIX
header is 20-24 bytes, depending on architecture. 128 bytes is used to
deal with an unlikely scenario where fields are added to the header.

A common value for net.core.wmem_max is 212992, which gives a max
message size of 212992 * 2 - 128 = 425856 bytes.

Very large net.core.wmem_max would yield very large xcm.max_msg_size
values. The XCM messaging API does not cater well for huge messages,
so UX and UXF caps their xcm.max_msg_size at 262144 bytes. In
addition, a very large xcm.max_msg_size may introduce issues for
applications using stack-allocated message buffers.

Prior max was hardcoded to 65535 bytes, with the assumption that
kernel runtime configuration would allow it.

The kernel may fail to allocate memory for large messages, in
particular on kernels using a single linear buffer to hold the message
payload.

The documentation is updated with a section on messaging transport
maximum message size. The current max is now spelled out in the
documentation, for all messaging transports.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>