Reduce control interface stack
In certain situations, libxcm control interface implementation used
~38 kb stack, which caused issues for threads with very small stacks.

With this change, the stack allocation is replaced with a heap
allocation, and is performed if (and only if) there seems to be a
message on the control interface socket.

This change also slightly improve performance, for the (rare) case
where a control client is connected for a long duration of time. For
other cases, the signaling performance remains unaffected.

Resolves #100

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>