Fix early DNS timeouts
A configured "dns.timeout" could finish early, in case the timeout
time was longer than <num-retries> * <per-retry-timeout> (both of
which are out of the control of the application).

The number of DNS query retries wasn't explicitly configured, and XCM
relied on the c-ares default.

With this change, the number of retries is deduced from the
user-configured timeout, and is explicitly configured in c-ares.

This bug was discovered as the default number of retries was reduced
in newer c-ares version (starting with 1.1.7), which broke the DNS
timeout test cases.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>