Add support for lists in attribute framework
Introduce support for attribute lists in the library-internal socket
attribute framework. This is a rewrite of this machinery.

The XCM attribute namespace was always hierachical, but the
implementation was a flat name -> <set of callbacks + meta data>
mapping. This worked well when attributes were organized into pseudo
dictionaries. It didn't work well with dynamically-sized lists of
attributes.

Instead of attribute being managed a flat name -> callback function
mapping, the attributes are now represented as a per-socket tree.

The tree has three types of nodes; dictionaries, lists and values. The
values are leafs, and dictionaries and lists are interior
nodes. Arbitrary nesting is allowed. The value nodes are either
boolean, integer, string, binary or double (per <xcm_attr_types.h>).

The attribute name syntax is extended to support lists, using square
brackets to indicate a list index (in an ALGOL 60-like manner).  For
example, a hypothetical fifth list element of a "somelist" key in the
"xcm" dictionary would have the name "xcm.somelist[4]". An attribute
name is a path specifying a node in the socket attribute tree.

A new library-internal type attr_path module is added, and which parse
names into a sequence of dictionary keys and list indices.

Some unspecified corner cases are now documented in the <xcm_attr.h>.
In particular, xcm_attr_set*() and xcm_attr_get*() are now documented
to set EINVAL in it's given a name with an invalid syntax.

This is consistent with how too-long names were already handled, but
could be viewed as an ABI breaking change, albeit unlikely to actually
break anything.

Attributes which values were lists coded as strings (e.g.,
"tls.peer_names") will remain so, to maintain backward compatibility.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>