Fix overly eager NUL termination
ut_aprintf() would attempt to NUL terminate the resulting buffer if
the vsnprintf() couldn't fit the result.

The NUL termination
1) was buggy and resulted in a far too short string.
2) was unnecessary, since the output from vsnprintf() is NUL terminated
   already.

This is not known to have caused any library-level issues. ut_aprintf()
is mostly used for creating trace messages, and the used buffers
are very rarely fully utilized (and thus not affected by this bug).

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>