Extend attribute API to accomodate list of attributes
The <xcm_attr.h> API is extended to support list nodes in the
attribute tree.

xcm_attr_get_list_len() is added to allow querying list length.

A set of xcm_attr_getf*() functions are added for convenient access to
attributes with dynamically created names (e.g., a particular element
in a list).

The API/ABI version is bumped to 26 and the library implementation
version to 1.11.0.

The attribute section in the <xcm.h> user manual is updated.

Signed-off-by: Mattias Rönnblom <mattias.ronnblom@ericsson.com>